Move integration method from Editor to Runtime.
