+++ Assets/VRM/UniVRM/Scripts/SkinnedMeshUtility/MeshIntegratorUtility.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: a982d9d30c0145038245b0214dc2f2e4
+timeCreated: 1560190306
